IPRT/posix/semwait.h: Build fix. bugref:10138
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@92781 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r3/posix/semwait.h
@@ -119,7 +119,7 @@ DECLINLINE(uint64_t) rtSemPosixCalcDeadline(uint32_t fFlags, uint64_t uTimeout, 
 
         pAbsDeadline->tv_sec  += TsAdd.tv_sec;
         pAbsDeadline->tv_nsec += TsAdd.tv_nsec;
-        if (pAbsDeadline->tv_nsec >= RT_NS_1SEC)
+        if ((uint32_t)pAbsDeadline->tv_nsec >= RT_NS_1SEC)
         {
             pAbsDeadline->tv_nsec -= RT_NS_1SEC;
             pAbsDeadline->tv_sec++;
